// src/src/core/constants/enums.ts
export enum PaymentStatus {
  Pending = 'pending',
  Completed = 'completed',
  Failed = 'failed',
  Canceled = 'canceled',
  Used = 'used',
}
// ...
export enum TransactionType {
  unit = 'unit',
  newLicense = 'new_license_payment',
  preRegistration = 'pre_registration_payment',
  licenseReplacement = 'license_replacement_payment',
  licenseRenewal = 'license_renewal_payment',
  drivingSchoolApplication = 'driving_school_application_payment',
  permitIssuance = 'permit_payment',
  biometricsPayment = 'biometrics_payment',
  cbtReschedulePayment = 'cbt_reschedule_payment',
  inspectionFee = 'inspection_payment',
  drivingSchoolCompletionPayment = 'driving_school_completion_payment',
  licenseRenewalMobile = 'license_renewal_payment_mobile',
}